version 0.9.0

+++ src/users/users.controller.ts
@@ -1,7 +1,8 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, BadRequestException } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { User } from './schema/user.schema';
 
 @Controller('users')
 export class UsersController {
@@ -13,22 +14,37 @@ export class UsersController {
   }
 
   @Get()
-  findAll() {
+  async findAll():Promise<User[]> {
     return this.usersService.findAll();
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  async findOne(@Param('id') id: string):Promise<User> {
     return this.usersService.findOne(id);
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto):Promise<User> {
+    //todo
     return this.usersService.update(id, updateUserDto);
   }
 
+  // @Delete(':id')
+  // async remove(@Param('id') id: string):Promise<User | unknown> {
+  //   return this.usersService.remove(id);
+  // }
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.usersService.remove(id);
+  async remove(@Param('id') id: string):Promise<any> {
+    if(id.length !=24) throw new BadRequestException({'message' : 'id debe que tener 24 caracteres'});
+
+    const eliminarUsuario = await this.usersService.remove(id);
+
+    if(!eliminarUsuario) throw new BadRequestException({'message': 'no existe este usuario'});
+
+  
+
+    return {'message':`el usuario con id ${id} ha borrado`};
   }
+  
+  
 }
